+++ tests/fixtures/smoke-tests/failing-test/tests/unit/some-test.js
@@ -1,7 +1,0 @@
-/*jshint strict:false */
-/* globals test, ok */
-
-test('failing test', function() {
-  ok(false, 'test should fail to confirm ember test exit code');
-});
-