+++ tests/helpers/mock-project.js
@@ -30,8 +30,4 @@ MockProject.prototype.name = function() {
   return 'mock-project';
 };
 
-MockProject.prototype.initializeAddons = function() {
-  this.addons = [];
-};
-
 module.exports = MockProject;